UI : Improved
Added :
- Display flex
- Content center
- Items center
