+++ static/js/main.js
@@ -5,7 +5,9 @@ import { SignUpElement } from "./elements/signup.js";
 import { NavBarElement } from "./elements/navbar.js";
 import { NavBarButtonElement } from "./elements/navbar-button.js";
 import { HeaderElement } from "./elements/header.js";
-import { CategoryButton } from "./elements/category-btn.js";
+import { CategoryButton } from "./elements/category.js";
+import { PostBox } from "./elements/post.js";
+import { CategoriesElement } from "./elements/categories.js";
 
 export let user = await User.createInstance();
 
@@ -17,4 +19,7 @@ customElements.define("c-signin", SignInElement);
 customElements.define("c-navbar", NavBarElement);
 customElements.define("c-navbutton", NavBarButtonElement);
 customElements.define("c-header", HeaderElement);
-customElements.define("c-categories", CategoryButton);
+customElements.define("c-categories", CategoriesElement);
+customElements.define("c-category", CategoryButton);
+
+customElements.define("c-postbox", PostBox);
